make tar package deal with bigger gid/uids
Port of fixes for https://issues.apache.org/jira/browse/COMPRESS-314 and
https://issues.apache.org/jira/browse/COMPRESS-315

+++ src/main/org/apache/tools/tar/TarInputStream.java
@@ -484,11 +484,11 @@ public class TarInputStream extends FilterInputStream {
             } else if ("linkpath".equals(key)){
                 currEntry.setLinkName(val);
             } else if ("gid".equals(key)){
-                currEntry.setGroupId(Integer.parseInt(val));
+                currEntry.setGroupId(Long.parseLong(val));
             } else if ("gname".equals(key)){
                 currEntry.setGroupName(val);
             } else if ("uid".equals(key)){
-                currEntry.setUserId(Integer.parseInt(val));
+                currEntry.setUserId(Long.parseLong(val));
             } else if ("uname".equals(key)){
                 currEntry.setUserName(val);
             } else if ("size".equals(key)){
